refactor: move Helpers.js -> _shared/helpers.js
