+++ packages/compile/src/ReadDat.js
@@ -4,7 +4,7 @@ import B from 'bufx'
 import byline from 'byline'
 import R from 'ramda'
 
-import { canonicalVensimName } from './Helpers.js'
+import { canonicalVensimName } from './_shared/helpers.js'
 
 /**
  * Read a Vensim `dat` file with static data and return a Map.
