+++ packages/compile/src/SubscriptRangeReader.js
@@ -4,7 +4,7 @@ import R from 'ramda'
 import XLSX from 'xlsx'
 import ModelReader from './ModelReader.js'
 import { Subscript } from './Subscript.js'
-import { cFunctionName, matchRegex, readCsv } from './Helpers.js'
+import { cFunctionName, matchRegex, readCsv } from './_shared/helpers.js'
 
 export default class SubscriptRangeReader extends ModelReader {
   constructor(modelDirname) {
