+++ packages/compile/src/VarNameReader.js
@@ -2,7 +2,7 @@ import ModelReader from './ModelReader.js'
 import R from 'ramda'
 import { createParser } from './Parser.js'
 import { sub, isIndex, normalizeSubscripts } from './Subscript.js'
-import { canonicalName } from './Helpers.js'
+import { canonicalName } from './_shared/helpers.js'
 //
 // VarNameReader reads a model var name using the parser to get the var name in C format.
 // This is used to generate a variable output in the output section.
