+++ packages/compile/src/index.js
@@ -2,7 +2,7 @@
 
 export { generateCode } from './CodeGen.js'
 export { parseAndGenerate, printNames } from './Entry.js'
-export { canonicalName } from './Helpers.js'
+export { canonicalName } from './_shared/helpers.js'
 export { parseModel } from './Parser.js'
 export { preprocessModel } from './Preprocessor.js'
 export { readDat } from './ReadDat.js'
